requirements.txt
# requirements.txt
psycopg2-binary==2.9.10
mysql-connector-python==8.1.0
python-dotenv==1.0.0